Use jumps instead of filling memory in stack overflow test
Behavior seemed weird with the memory fill method, the processor would
read an invalid zero instruction after a bunch of pushes.

The test did continue to pass, but not because the processor ran into a
stack overflow condition. This way ensures execution can only ever fail
due to a stack overflow.